+++ src/types/session_data.rs
@@ -1,0 +1,13 @@
+use crate::types::QuotedString;
+
+/// Session data.
+///
+/// See: [4.3.4.4. EXT-X-SESSION-DATA]
+///
+/// [4.3.4.4. EXT-X-SESSION-DATA]: https://tools.ietf.org/html/rfc8216#section-4.3.4.4
+#[allow(missing_docs)]
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub enum SessionData {
+    Value(QuotedString),
+    Uri(QuotedString),
+}
